Polynomial scaling factor added to CableComponent.
Reason:
The OTLS models calculation libraries attempt to use consistent units throughout
so unit conversions do not have to be used deep within the classes. Makes it way
more simple when using multiple unit systems (metric/imperical). The polynomials
were not adhering to this because the calculation classes were expecting percent
strain and load (percent strain being the problem). For the code that depends on
the Models library, it is not feasible to (a) redefine the polynomials in strain
units or (b) reduce all loads by a factor of 100. The scaling factor allows
inconsistent units to be used for the polynomial, but scaled to nominal
(strain or load) values in the calculation classes.

Changes:
- CableComponent
  - scale_polynomial class members added

- SagTensionCableComponent
  - Members exposed with getters/setters
  - Validation updated

- CableComponentElongationModel
  - Hardcoded 100 reduction removed and replaced with scaling factors